Refactor chat-ops to its own directory
