+++ __tests__/operations/chat-ops/assign.test.ts
@@ -1,5 +1,5 @@
-import assign from '../../src/operations/assign'
-import {Command, Commands} from "../../src/command";
+import assign from '../../../src/operations/chat-ops/assign'
+import {Command, Commands} from "../../../src/command";
 import * as github from "@actions/github";
 import * as core from "@actions/core";
 import nock from "nock";
