+++ __tests__/operations/chat-ops/close.test.ts
@@ -1,5 +1,5 @@
-import close from '../../src/operations/close'
-import {Command, Commands} from "../../src/command";
+import close from '../../../src/operations/chat-ops/close'
+import {Command, Commands} from "../../../src/command";
 import * as github from "@actions/github";
 import * as core from "@actions/core";
 import nock from "nock";
