+++ __tests__/operations/chat-ops/comment.test.ts
@@ -1,5 +1,5 @@
-import comment from '../../src/operations/comment'
-import {Command, Commands} from "../../src/command";
+import comment from '../../../src/operations/chat-ops/comment'
+import {Command, Commands} from "../../../src/command";
 import * as github from "@actions/github";
 import * as core from "@actions/core";
 import nock from "nock";
