+++ __tests__/operations/chat-ops/label.test.ts
@@ -1,9 +1,9 @@
-import labelChatOps from '../../src/operations/label-chat-ops'
-import {Command, Commands} from "../../src/command";
+import labelChatOps from '../../../src/operations/chat-ops/label'
+import {Command, Commands} from "../../../src/command";
 import * as github from "@actions/github";
 import * as core from "@actions/core";
 import nock from "nock";
-import {addLabels} from "../../src/github";
+import {addLabels} from "../../../src/github";
 
 const postLabels = jest.fn()
 const deleteLabels = jest.fn()
