+++ src/operations/chat-ops/assign.ts
@@ -1,6 +1,6 @@
-import {ChatOps} from '../config'
-import {Commands} from '../command'
-import {assign} from '../github'
+import {ChatOps} from '../../config'
+import {Commands} from '../../command'
+import {assign} from '../../github'
 
 export default async function (
   chatOps: ChatOps,
